requirements.txt
awesome-slugify==1.6.5
dataset==0.7.1
fuzzywuzzy==0.15.0
geocoder==1.19.0
ipdb==0.10.2
ipykernel==4.6.1
ipython==6.0.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
jupyter==1.0.0
jupyter-client==5.0.1
jupyter-console==5.1.0
jupyter-core==4.3.0
matplotlib==2.0.0
numpy==1.12.1
pandas==0.20.1
pandas-linker
PyPrind==2.10.0
python-dateutil==2.6.0
python-Levenshtein==0.12.0
pytz==2017.2
xlrd==1.0.0
xlwt==1.2.0
